Extract RN Fabric host config
